More big stuff
I keep moving around and renaming a lot of files so it looks like I do more than I actually do haha.
Created FileIdInterpreter to store a file interpreting system for conciseness. TODO generalise all board entities under BoardEntity, if it will help (may not, with the file system in place).
Created LevelInfo as a tracker for all the information in a level such as tile/entity positions. Pasted a bunch of code from WorldGenInterpreter (now obsolete there). Haven't cleaned it up yet, so it's not functional. I seem to never commit on a finished note, but rather halfway through an undertaking.
Created portal objects, and a new portal_id file. This makes moving between levels much easier to code. Everything I do is to make the overall project easier to manage.